Revert "Revert "new utils""
This reverts commit 629b21c0627454cbba5154a2cc05ef10d77bcd48.
