+++ mlrun/api/crud/client_spec.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-import mlrun.api.crud.runtimes.nuclio.helpers
+import mlrun.api.utils.runtimes.nuclio
 import mlrun.common.schemas
 import mlrun.utils.singleton
 from mlrun.config import Config, config, default_config
@@ -45,7 +45,7 @@ class ClientSpec(
                 config.dask_kfp_image, client_version, client_python_version
             ),
             api_url=config.httpdb.api_url,
-            nuclio_version=mlrun.api.crud.runtimes.nuclio.helpers.resolve_nuclio_version(),
+            nuclio_version=mlrun.api.utils.runtimes.nuclio.resolve_nuclio_version(),
             spark_operator_version=config.spark_operator_version,
             calculate_artifact_hash=config.artifacts.calculate_hash,
             generate_artifact_target_path_from_artifact_hash=config.artifacts.generate_target_path_from_artifact_hash,
